Handle story deep link on home screen
